Video modes and SSE checking
Implemented the foundations for a video driver, this will provide methods for
putting information onto the display no matter the graphical mode booted with.

In implementing the graphical foundations, an optimisation the compiler uses is
to use the SSE (xmm_) registers, these are non-standard registers which the
bootloader has to test for and enable if they exist else they cannot be used.
Compiler options were added in the Makefile to disable their use.